Lots of small cleanups from vdr-1.6 ebuilds. (Portage version: 2.1.6.13/svn/Linux i686)
svn path=/vdr-devel/; revision=605

+++ media-video/vdr/vdr-1.7.0-r1.ebuild
@@ -32,8 +32,6 @@ EXT_P=VDR-Extensions-Patch-${EXT_V}
 EXT_DIR=${WORKDIR}/${EXT_P}/
 EXT_VDR_PV="${PV/_p/-}"
 
-APIWRAPPER=vdr-1.7.0-multiproto-update-dvb-api-wrapper-0.4.diff
-
 DESCRIPTION="Video Disk Recorder - turns a pc into a powerful set top box for DVB"
 HOMEPAGE="http://www.cadsoft.de/vdr/"
 SRC_URI="ftp://ftp.cadsoft.de/vdr/Developer/${MY_P}.tar.bz2
@@ -181,9 +179,6 @@ src_unpack() {
 	unpack ${A}
 
 	cd "${S}"
-	# fix for wrong header include #263840 ; this need >libdvdread-0.9.7
-	sed -e "s:dvdread:dvdnav:g" -i "${EXT_DIR}/${P}"_extensions.diff
-	
 	#applying maintainace-patches
 
 	ebegin "Changing pathes for gentoo"
@@ -237,6 +232,7 @@ src_unpack() {
 	eend 0
 
 	epatch "${FILESDIR}"/vdr-1.6.0-makefile-install-header.diff
+
 	if use dxr3; then
 		einfo "Applying dxr3 subtitle hack"
 		epatch "${FILESDIR}"/vdr-1.6.0-dxr3-subtitlehack.diff
@@ -254,8 +250,12 @@ src_unpack() {
 
 		cd "${S}"
 		# Now apply extensions patch
-		local fname="${PN}-${EXT_VDR_PV:-${PV}}_extensions.diff"
-		epatch "${EXT_DIR}/${fname}"
+		local fname="${EXT_DIR}/${PN}-${EXT_VDR_PV:-${PV}}_extensions.diff"
+
+		# fix for wrong header include #263840 ; this need >libdvdread-0.9.7
+		sed -e "s:dvdread:dvdnav:g" -i "${fname}"
+
+		epatch "${fname}"
 
 		if use h264; then
 			epatch "${EXT_DIR}/extras/vdr-1.7.0-ext_h264.diff"
@@ -304,6 +304,8 @@ src_unpack() {
 		for flag in $EXT_PATCH_FLAGS; do
 			use $flag && enable_patch ${flag}
 		done
+
+		# patches that got renamed
 		use iptv && enable_patch pluginparam
 		use liemikuutio && enable_patch liemiext
 		eend 0
@@ -314,7 +316,7 @@ src_unpack() {
 		emake .dependencies >/dev/null
 		eend $? "make depend failed"
 
-		do_unifdef
+		[[ -z "$NO_UNIFDEF" ]] && do_unifdef
 
 		if use ehd; then
 			einfo "Applying Reel eHD extension patch"
@@ -381,7 +383,7 @@ src_install() {
 	local f
 	rm *vdr-1.4* 2>/dev/null
 	for f in *; do
-		[[ -f ${f} ]] || break
+		[[ -f ${f} ]] || continue
 		newdoc "${f}" "${f}".ExtensionsPatch || die "Could not install extensions-patch doc ${f}"
 	done
 
@@ -405,9 +407,17 @@ src_install() {
 	chown -R vdr:vdr "${D}/${CONF_DIR}"
 }
 
+pkg_preinst() {
+	has_version "<${CATEGORY}/${PN}-1.3.36-r3"
+	previous_less_than_1_3_36_r3=$?
+
+	has_version "<${CATEGORY}/${PN}-1.6.0"
+	previous_less_than_1_6_0=$?
+}
+
 pkg_postinst() {
 	elog "It is a good idea to run vdrplugin-rebuild now."
-	if has_version "<media-video/vdr-1.3.36-r3"; then
+	if [[ $previous_less_than_1_3_36_r3 = 0 ]] ; then
 		ewarn "Upgrade Info:"
 		ewarn
 		ewarn "If you had used the use-flags lirc, rcu or vfat"
@@ -453,6 +463,17 @@ pkg_postinst() {
 		ewarn "ATSC is only supported by a rudimentary patch"
 		einfo "and need at least this patch and a plugin installed"
 		einfo "emerge media-plugins/vdr-atscepg"
+	fi
+
+	if [[ $previous_less_than_1_6_0 = 0 ]]; then
+		elog "By default vdr is now started with utf8 character encoding"
+		elog
+		elog "To rename the old recordings to utf8 conforming names, do this:"
+		elog "\temerge app-text/convmv"
+		elog "\tconvmv -f latin1 -t utf8 -r --notest -i /var/vdr/video/"
+		elog
+		elog "To fix the descriptions of your recordings do this:"
+		elog "\tfind /var/vdr/video/ -name "info.vdr" -print0|xargs -0 recode latin1..utf8"
 	fi
 
 	elog "To get nice symbols in OSD we recommend to install"
