+++ media-video/vdr/vdr-1.7.1.ebuild
@@ -1,4 +1,4 @@
-# Copyright 1999-2008 Gentoo Foundation
+# Copyright 1999-2009 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
 # $Header: /var/cvsroot/gentoo-x86/media-video/vdr/vdr-1.6.0_p1.ebuild,v 1.6 2008/05/21 05:50:50 zzam Exp $
 
@@ -223,6 +223,7 @@ src_unpack() {
 	eend 0
 
 	epatch "${FILESDIR}"/vdr-1.6.0-makefile-install-header.diff
+
 	if use dxr3; then
 		einfo "Applying dxr3 subtitle hack"
 		epatch "${FILESDIR}"/vdr-1.6.0-dxr3-subtitlehack.diff
@@ -240,8 +241,12 @@ src_unpack() {
 
 		cd "${S}"
 		# Now apply extensions patch
-		local fname="${PN}-${EXT_VDR_PV:-${PV}}_extensions.diff"
-		epatch "${EXT_DIR}/${fname}"
+		local fname="${EXT_DIR}/${PN}-${EXT_VDR_PV:-${PV}}_extensions.diff"
+
+		# fix for wrong header include #263840 ; this need >libdvdread-0.9.7
+		sed -e "s:dvdread:dvdnav:g" -i "${fname}"
+
+		epatch "${fname}"
 
 		# Fix typo in Make.config.template
 		sed -e 's/CMDRECMDI18N/CMDRECCMDI18N/' -i Make.config.template
@@ -277,7 +282,8 @@ src_unpack() {
 			local IGNORE_PATCHES="channelscan pluginapi pluginmissing"
 
 			extensions_all_defines > "${T}"/new.IUSE
-			echo $EXT_PATCH_FLAGS $IGNORE_PATCHES|tr ' ' '\n' |sort > "${T}"/old.IUSE
+			echo $EXT_PATCH_FLAGS $IGNORE_PATCHES | \
+				tr ' ' '\n' |sort > "${T}"/old.IUSE
 			local DIFFS=$(diff -u "${T}"/old.IUSE "${T}"/new.IUSE|grep '^[+-][^+-]')
 			if [[ -z ${DIFFS} ]]; then
 				einfo "EXT_PATCH_FLAGS is up to date."
@@ -303,7 +309,7 @@ src_unpack() {
 		emake .dependencies >/dev/null
 		eend $? "make depend failed"
 
-		do_unifdef
+		[[ -z "$NO_UNIFDEF" ]] && do_unifdef
 
 		use iptv && sed -i sources.conf -e 's/^#P/P/'
 	fi
@@ -365,7 +371,7 @@ src_install() {
 	local f
 	rm *vdr-1.4* 2>/dev/null
 	for f in *; do
-		[[ -f ${f} ]] || break
+		[[ -f ${f} ]] || continue
 		newdoc "${f}" "${f}".ExtensionsPatch || die "Could not install extensions-patch doc ${f}"
 	done
 
@@ -389,6 +395,14 @@ src_install() {
 	chown -R vdr:vdr "${D}/${CONF_DIR}"
 }
 
+pkg_preinst() {
+	has_version "<${CATEGORY}/${PN}-1.3.36-r3"
+	previous_less_than_1_3_36_r3=$?
+
+	has_version "<${CATEGORY}/${PN}-1.6.0"
+	previous_less_than_1_6_0=$?
+}
+
 pkg_postinst() {
 	elog "!!!! WARNING !!!!"
 	elog "  ${P} contains large changes with respect to"
@@ -400,7 +414,7 @@ pkg_postinst() {
 	elog
 	elog
 	elog "It is a good idea to run vdrplugin-rebuild now."
-	if has_version "<media-video/vdr-1.3.36-r3"; then
+	if [[ $previous_less_than_1_3_36_r3 = 0 ]] ; then
 		ewarn "Upgrade Info:"
 		ewarn
 		ewarn "If you had used the use-flags lirc, rcu or vfat"
@@ -446,6 +460,17 @@ pkg_postinst() {
 		ewarn "ATSC is only supported by a rudimentary patch"
 		einfo "and need at least this patch and a plugin installed"
 		einfo "emerge media-plugins/vdr-atscepg"
+	fi
+
+	if [[ $previous_less_than_1_6_0 = 0 ]]; then
+		elog "By default vdr is now started with utf8 character encoding"
+		elog
+		elog "To rename the old recordings to utf8 conforming names, do this:"
+		elog "\temerge app-text/convmv"
+		elog "\tconvmv -f latin1 -t utf8 -r --notest -i /var/vdr/video/"
+		elog
+		elog "To fix the descriptions of your recordings do this:"
+		elog "\tfind /var/vdr/video/ -name "info.vdr" -print0|xargs -0 recode latin1..utf8"
 	fi
 
 	elog "To get nice symbols in OSD we recommend to install"
